fix route, model, view

+++ database/migrations/2024_11_29_051740_create_gudangs_table.php
@@ -13,8 +13,11 @@ return new class extends Migration
     {
         Schema::create('gudangs', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('id_user');
-            $table->foreign('id_user')->references('id')->on('users');
+            $table->unsignedBigInteger('id_user'); // Tipe data disesuaikan
+            $table->foreign('id_user')
+                  ->references('id')
+                  ->on('users');
+                //   ->onDelete('cascade'); // Opsional
             $table->integer('kapasitas');
             $table->text('deskripsi');
             $table->timestamps();
